requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
python-decouple>=3.8,<4.0
anthropic>=0.39,<1.0
uvicorn>=0.32,<1.0
psycopg2-binary>=2.9,<3.0
